+++ tests/integration/stack_tags/program/index.ts
@@ -1,0 +1,7 @@
+// Copyright 2016-2025, Pulumi Corporation.  All rights reserved.
+
+import * as pulumi from "@pulumi/pulumi";
+
+import { Resource } from "./resource";
+
+new Resource("hello", {});